修复输入的保质期为0的时候，或者输入已经存在的保质期或种类的时候会crash的问题

+++ app/src/main/java/lying/fengfeng/foodrecords/utils/DateUtil.kt
@@ -40,7 +40,7 @@ object DateUtil {
         } catch (e: ParseException) {
             try {
                 val productionTimeMillis = dateFormatter.parse(productionDate).time
-                val expirationTimeMillis = productionTimeMillis + shelfLife.toInt() * (24 * 60 * 60 * 1000)
+                val expirationTimeMillis = productionTimeMillis + shelfLife.toLong() * (24 * 60 * 60 * 1000)
                 ((expirationTimeMillis - System.currentTimeMillis()) / (24 * 60 * 60 * 1000)).toInt()
             } catch (e: ParseException) {
                 0
